interfaces moved to models
